added routing split market place app into different views
