+++ example/store.z/src/package.json
@@ -6,12 +6,16 @@
   "scripts": {
     "clear": "rm -rf ../views/* .cache .parcel-cache",
     "start": "parcel src/index.zhtml",
-    "build": "npm run clear && cp -R src/images ../views/. && parcel build src/index.html -d ../views --out-file ../views/index.zhtml --no-source-maps",
-    "watch": "npm run clear && cp -R src/images ../views/. && parcel watch src/index.html -d ../views --out-file ../views/index.zhtml",
+    "build": "npm run clear && cp -R src/images ../views/. && parcel build src/index.html -d ../views --out-file --no-source-maps",
+    "watch": "npm run clear && cp -R src/images ../views/. && parcel watch src/index.html -d ../views",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "Point Network",
   "license": "ISC",
+  "alias": {
+    "~": "./src",
+    "Common": "./src/Common"
+  },
   "devDependencies": {
     "@types/node": "^15.0.2",
     "@types/react": "^17.0.5",
@@ -21,6 +25,7 @@
   },
   "dependencies": {
     "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react-dom": "^17.0.2",
+    "wouter": "^2.7.4"
   }
 }
