+++ example/store.z/src/src/Products/index.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export const Products = () => {
+    return (
+        <h2>Products:...</h2>
+    )
+}
